feat(WIP): Still working on policyId fetching
Set up "<Restriction>policies" endpoints in AccessPolicyController instead, as a central location
