+++ service/src/main/groovy/com/k_int/accesscontrol/grails/AccessControlUrlMapping.groovy
@@ -8,7 +8,6 @@ class AccessControlUrlMapping {
         String controller = mapping.controller
 
         "${path}/canCreate"(controller: controller, action: "canCreate", method: 'GET')
-        "${path}/canClaim"(controller: controller, action: "canClaim", method: 'GET')
 
         "${path}/$id/canRead"(controller: controller, action: "canRead", method: 'GET')
         "${path}/$id/canUpdate"(controller: controller, action: "canUpdate", method: 'GET')
